Improved comments, moved method to intuitive location, fixed imports.
