+++ fsc_analyze.py
@@ -22,6 +22,8 @@ import matplotlib
 matplotlib.use('TkAgg')
 import matplotlib.pyplot as plt
 
+from fsc_launch import INPUT_DATA_CSV
+
 from utils import read_csv_file, extract_data_from_dataframe, extract_data_from_csv, \
 	extract_data_for_date_from_dataframe
 
@@ -31,8 +33,8 @@ def find_worst_results(filename, num_worst=5):
 	dictionary of length num_worst where the key is the disagreement rate and the value is the timestamp."""
 	frame = read_csv_file(filename)
 	net_times = set(extract_data_from_dataframe(frame, "timestamp_utc"))
-	shcu = read_csv_file('shcu_good_data.csv')
-	shcu_times = set(extract_data_from_csv('shcu_good_data.csv', "timestamp_utc"))
+	shcu = read_csv_file(INPUT_DATA_CSV)
+	shcu_times = set(extract_data_from_dataframe(shcu, "timestamp_utc"))
 	times = net_times.intersection(shcu_times)
 	disagreement_rates = [(-1, '')] * num_worst
 	heapq.heapify(disagreement_rates)
