+++ preprocess_stamps_launch.py
@@ -8,8 +8,24 @@ Recommended to run this script on blt with the command:
 SGE_Batch -r "preprocess_stamps" -c "python3 -u preprocess_stamps_launch.py" -P 1
 """
 
-from preprocess_setup_launch import OUTPUT_DIR, RES_DIR, create_constant_mask
-from utils import extract_times_from_files_in_directory, separate_data, BLACK
+import os
+import numpy as np
+from PIL import Image
+from scipy import misc
+
+from preprocess_setup_launch import OUTPUT_DIR, RES_DIR
+from utils import extract_times_from_files_in_directory, separate_data, BLACK, BLUE
+
+
+def create_constant_mask(color, filename):
+	"""Creates a mask where any pixels not always of color are BLUE. Saves it in filename."""
+	b_mask = np.full((480, 480, 3), color)
+	for dirpath, subdirs, files in os.walk(OUTPUT_DIR + '/simplemask/'):
+		for file in files:
+			img = misc.imread(os.path.join(dirpath, file))
+			b_mask[(img != color).any(axis=2)] = BLUE
+	Image.fromarray(b_mask.astype('uint8')).save(OUTPUT_DIR + '/' + filename)
+
 
 if __name__ == "__main__":
 	times = extract_times_from_files_in_directory(RES_DIR)
