+++ process_launch.py
@@ -1,5 +1,7 @@
 """
-Call this program after completing all preprocess launch tasks. This method goes through all cropped sky images and
+This program can be called after the train_launch.py finishes its task.
+
+This method goes through all cropped sky images and
 produces the network output mask for each. Breaks up tasks into many smaller batches to improve speed. Specify the
 number of batches per network to increase the number of batches created & increase speed. Make sure that the number
 of batches times the number of batches per network is small enough so that BLT can fit all of the jobs.
@@ -11,7 +13,7 @@ import os
 INPUT_DIR = "bad_data"
 
 # Specify the location of the csv file that contains desired timestamp_utc information.
-input_csv = "shcu_bad_data.csv"
+INPUT_DATA_CSV = "shcu_bad_data.csv"
 
 # Specify the base of the job id
 JOB_NAME = "bad-net-"
@@ -26,14 +28,14 @@ if __name__ == "__main__":
 	num_batches = len(exp_labels)
 
 	total_length = -1  # This file has a header
-	for line in open(input_csv):
+	for line in open(INPUT_DATA_CSV):
 		total_length += 1
 
 	batch_length = int(total_length / num_batches_per_network)
 
 	for exp_label in exp_labels:
 		for i in range(num_batches_per_network):
-			name = JOB_NAME + exp_label + "-" + str(i)
+			name = JOB_NAME + exp_label + '-{:0>2}'.format(i)
 			start = batch_length * i
 			finish = batch_length * (i + 1) if batch_length * (i + 1) < total_length else total_length
 			os.system('SGE_Batch -r "{}" -c "python3 -u process.py {} {} {}" -P 1'.format(name, exp_label, int(start),
